# Optimize multiple integer array indexing for cupy.ndarray.__getitem__