Merge after #2126.

This PR optimizes `cupy.ndarray.__getitem__` by removing unnecessary copy call using a trick descrbied in #2126.

Here is a benchmark code.

```python
# getitem_benchmark.py
#
# on TitanX (Pascal), CUDA 8.0:
# optimized:  2.8198 (ms)
# unoptimized 3.8088 (ms)

import cupy

def benchmark(a, slices):
    start = cupy.cuda.Event()
    end = cupy.cuda.Event()
    start.record()
    a[slices]
    end.record()
    end.synchronize()
    time = cupy.cuda.get_elapsed_time(start, end)
    return time

n = 3000
slices = (cupy.arange(n), slice(None), cupy.arange(n))
a = cupy.ones((n, 2, n))
cupy.cuda.Stream(null=True).synchronize()

# dry run
for i in range(5):
    benchmark(a, slices)
with cupy.cuda.profile():
    print benchmark(a, slices)
```

And result of kernel calls.
Before optimization
```
 97.75%  809.65us         1  809.65us  809.65us  809.65us  cupy_copy
  0.62%  5.1200us         1  5.1200us  5.1200us  5.1200us  cupy_take_0axis
  0.53%  4.3520us         2  2.1760us  1.8240us  2.5280us  cupy_remainder
  0.45%  3.7440us         3  1.2480us  1.1200us  1.5040us  [CUDA memcpy DtoD]
  0.33%  2.7520us         2  1.3760us  1.1840us  1.5680us  cupy_multiply
  0.32%  2.6880us         1  2.6880us  2.6880us  2.6880us  cupy_sum
```
After optimization
```
26.51%  3.6480us         1  3.6480us  3.6480us  3.6480us  cupy_take
22.79%  3.1360us         2  1.5680us  1.4720us  1.6640us  cupy_remainder
18.60%  2.5600us         2  1.2800us  1.2160us  1.3440us  cupy_multiply
17.21%  2.3680us         2  1.1840us  1.1200us  1.2480us  [CUDA memcpy DtoD]
14.88%  2.0480us         1  2.0480us  2.0480us  2.0480us  cupy_sum
```