@rezoo 2017-02-02
Thank you for your PR. Could you resolve the conflicts?