@rezoo 2017-02-06
I see. In order to avoid unnecessary copy you implemented additional arguments in `_take` function. LGTM.